Fix Issue with Library Mutating Input Param Object.
This commit fixes an issue where the input params objects were being
mutated.

It seems quite unexpected that the provided input params would be
mutated by the stream.

This was causing issues when using a list of headers that is
a subset of the whole CSV column set as therefore desiring to also
apply the same header set to the includeColumns so that resulting
JSON object doesn't have extra properties; Field1, Field2... etc.